+++ controllers/api/notification.api.js
@@ -1,17 +1,17 @@
-var { addressModel, userModel } = require('../models/bookStore.model');
+var { notificationModel, userModel } = require('../../models/bookStore.model');
 
 var objReturn = {
     status: 1,
     msg: 'ok'
 }
 
-exports.getListAddress = async (req, res, next) => {
-    let listAddress = [];
+exports.getListNotification = async (req, res, next) => {
+    let listNotification = [];
 
     try {
-        listAddress = await addressModel.find();
-        if (listAddress) {
-            objReturn.data = listAddress;
+        listNotification = await notificationModel.find();
+        if (listNotification) {
+            objReturn.data = listNotification;
             objReturn.status = 1;
             objReturn.msg = 'lay thanh cong'
 
@@ -28,14 +28,14 @@ exports.getListAddress = async (req, res, next) => {
     return res.json(objReturn);
 }
 
-exports.getOneAddress = async (req, res, next) => {
+exports.getOneNotification = async (req, res, next) => {
     try {
         const id = req.params.id;
 
-        const address = await addressModel.findById(id);
+        const notification = await notificationModel.findById(id);
 
-        if (address) {
-            objReturn.data = address;
+        if (notification) {
+            objReturn.data = notification;
             objReturn.status = 1;
             objReturn.msg = 'lay thanh cong'
 
@@ -52,19 +52,18 @@ exports.getOneAddress = async (req, res, next) => {
     return res.json(objReturn);
 }
 
-exports.addAddress = async (req, res, next) => {
+exports.addNotification = async (req, res, next) => {
 
     try {
         const newData = req.body;
-        const checkUsername = await addressModel.findOne({ username: req.body.username })
-        const checkPhone = await addressModel.findOne({ phone: req.body.phone })
-        const checkLocation = await addressModel.findOne({ location: req.body.location })
+        const checkContent = await notificationModel.findOne({ content: req.body.content });
+        const checkTitle = await notificationModel.findOne({ title: req.body.title });
 
-        if ((checkUsername === null) && (checkPhone === null) && (checkLocation === null)) {
-            const dataRes = await addressModel.create(newData);
+        if ((checkTitle === null) && (checkContent === null)) {
+            const dataRes = await notificationModel.create(newData);
             return res.status(200).json({ message: dataRes })
         } else {
-            return res.status(400).json({ message: "Address da ton tai" })
+            return res.status(400).json({ message: "Thong bao da ton tai" })
         }
     } catch (error) {
         objReturn.status = 0;
@@ -78,7 +77,7 @@ exports.updateAddress = async (req, res, next) => {
     try {
         const id = req.params.id;
         const updatedData = req.body;
-        const result = await addressModel.findByIdAndUpdate(id, updatedData, { new: true });
+        const result = await notificationModel.findByIdAndUpdate(id, updatedData, { new: true });
 
         return res.status(200).json(result);
 
@@ -90,14 +89,14 @@ exports.updateAddress = async (req, res, next) => {
     return res.json(objReturn);
 }
 
-exports.deleteAddress = async (req, res, next) => {
+exports.deleteNotification = async (req, res, next) => {
 
     try {
         const id = req.params.id;
 
-        await userModel.deleteMany({ address: id });
+        await userModel.deleteMany({ notifications: id });
 
-        const result = await addressModel.findByIdAndDelete(id);
+        const result = await notificationModel.findByIdAndDelete(id);
         return res.status(200).json(result);
     } catch (error) {
         objReturn.status = 0;
